+++ src/main/java/com/dsip/backend/exception/InvalidCapitalUpdateException.java
@@ -6,9 +6,9 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 @ResponseStatus(HttpStatus.BAD_REQUEST)
 public class InvalidCapitalUpdateException extends DsipException {
 
-    public InvalidCapitalUpdateException(Integer currentAmount, Integer requestedAmount) {
+    public InvalidCapitalUpdateException(Double currentAmount, Double requestedAmount) {
         super(String.format(
-                "Total capital planned cannot be decreased. Current: %d, Requested: %d",
+                "Total capital planned cannot be decreased. Current: %.2f, Requested: %.2f",
                 currentAmount, requestedAmount));
     }
 }
